requirements.txt
pip~=25.0.1
pillow~=10.4.0
Django~=5.1.6
djangorestframework~=3.15.2